+++ pygpsclient/helpers.py
@@ -15,7 +15,7 @@ Created on 17 Apr 2021
 import re
 from subprocess import run as subrun
 from sys import executable
-from tkinter import Toplevel, Label, Button, W
+from tkinter import Toplevel, Entry, Label, Button, W
 from math import sin, cos, pi
 from datetime import datetime, timedelta
 from pygpsclient.globals import (
@@ -23,17 +23,18 @@ from pygpsclient.globals import (
     DEVICE_ACCURACY,
     HDOP_RATIO,
     FIXLOOKUP,
+    ENTCOL,
+    ERRCOL,
 )
 
-URLREGEX = re.compile(
-    # r"^(?:http|https)?://"  # http:// or https://
-    r"^(?:(?:[A-Z0-9](?:[A-Z0-9-]{0,61}[A-Z0-9])?\.)+(?:[A-Z]{2,6}\.?|[A-Z0-9-]{2,}\.?)|"  # domain...
-    r"localhost|"  # localhost...
-    r"\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})"  # ...or ip
-    r"(?::\d+)?"  # optional port
-    r"(?:/?|[/?]\S+)$",
-    re.IGNORECASE,
-)
+# validation type flags
+MAXPORT = 65535
+MAXALT = 10000.0  # meters arbitrary
+VALBLANK = 1
+VALNONBLANK = 2
+VALINT = 4
+VALFLOAT = 8
+VALURL = 16
 
 
 class ConfirmBox(Toplevel):
@@ -501,7 +502,7 @@ def corrage2int(code: int) -> int:
     :rtype: int
     """
 
-    LOOKUP = {
+    lookup = {
         0: 0,
         1: 1,
         2: 2,
@@ -516,7 +517,7 @@ def corrage2int(code: int) -> int:
         11: 120,
     }
 
-    return LOOKUP.get(code, 0)
+    return lookup.get(code, 0)
 
 
 def validURL(url: str) -> bool:
@@ -527,5 +528,48 @@ def validURL(url: str) -> bool:
     :return: valid True/False
     :rtype: bool
     """
+    # pylint: disable=line-too-long
 
-    return re.match(URLREGEX, url) is not None
+    regex = re.compile(
+        # r"^(?:http|https)?://"  # http:// or https://
+        r"^(?:(?:[A-Z0-9](?:[A-Z0-9-]{0,61}[A-Z0-9])?\.)+(?:[A-Z]{2,6}\.?|[A-Z0-9-]{2,}\.?)|"  # domain...
+        r"localhost|"  # localhost...
+        r"\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})"  # ...or ip
+        r"(?::\d+)?"  # optional port
+        r"(?:/?|[/?]\S+)$",
+        re.IGNORECASE,
+    )
+
+    return re.match(regex, url) is not None
+
+
+def valid_entry(entry: Entry, valmode: int, low=None, high=None) -> bool:
+    """
+    Validates tkinter entry field and highlights it if in error.
+
+    :param Entry entry: tkinter entry widget
+    :param int valmode: int representing validation type - can be OR'd
+    :param object low: optional min value
+    :param object high: optional max value
+    :return: True/False
+    :rtype: bool
+    """
+
+    valid = True
+    try:
+        if valmode & VALBLANK and entry.get() == "":  # blank ok
+            valid = True
+        else:
+            if valmode & VALNONBLANK:  # non-blank
+                valid = entry.get() != ""
+            if valmode & VALINT:  # int in range
+                valid = low < int(entry.get()) < high
+            if valmode & VALFLOAT:  # float in range
+                valid = low < float(entry.get()) < high
+            if valmode & VALURL:  # valid URL
+                valid = validURL(entry.get())
+    except ValueError:
+        valid = False
+
+    entry.config(bg=ENTCOL if valid else ERRCOL)
+    return valid